tedxlublin/requirements.txt
django
gunicorn
<<<<<<< HEAD
django-environ
=======
django-environ
airtable-python-wrapper
>>>>>>> 4d329a1 (Dodane podstrony /newsletter /game m.in.)
